requirements.txt
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
sendgrid==6.11.0
python-dotenv==1.0.1
certifi==2024.2.2 